test: HttpSessionAuthentification 테스트 코드 추가
